Moved LevelData to it's own file
better in terms of organization because both levelplayer and leveleditor use a leveldata object to load levels
